formula_support: move KegOnlyReason to /extend/os/mac
