+++ Library/Homebrew/extend/os/formula_support.rb
@@ -1,0 +1,5 @@
+require "formula_support"
+
+if OS.mac?
+  require "extend/os/mac/formula_support"
+end
